Set watcher `id` in child spec
What changed?
=============

Since tests can navigate through several LiveViews, and we want to
create a new watcher for every LiveView (given the current design of the
`LiveViewWatcher`), we update the call to `start_supervised` so it
passes a different `id` for each LiveView (which will pass it to
`start_child`). Otherwise, we're attempting to start the same supervised
process (hence the previous match on `:already_started`).

We could've chosen many things for the `id`, and it's common to define
it as an atom, but it just needs to be an erlang term. So, we make it
the `view.pid` -- simple (since we already have it) and unique enough
for our purposes.