Catch exits when checking async pids
What changed?
=============

There are scenarios when checking for async pids is too slow and the
LiveView is already dead by the time we call `assert_has` (which, in
turn, calls `watch_view`).

I don't have a lot of data on that, but anecdotally, I think
that can happen if a LiveView's `on_mount` callback is performing some
async tasks.

So, we reintroduce some of the `catch` code that LiveViewTest does
around calling functions like this. Honestly, it should've always been
there. So, if we catch the exit due to a redirect or live_redirect, we
can notify the LiveView of that happening.