Allow watcher to watch multiple views
What changed?
=============

It's possible a Watcher may need to watch multiple LiveViews. To support
that, we update the `LiveViewWatcher` code to start by monitoring the
original LiveView, but to also take in a `view` in `watch_view` so that
we can watch multiple LiveViews.

The watcher now sends messages back to the test with the watched view's
`pid` for pattern matching. We considered using `refs`, but didn't seem
like we needed the uniqueness and extra passing around of values for
that. Since this is happening in tests that exit quickly, the `pid`
seems fine.