Watch sequential LiveViews (via redirects) and async behavior
What changed?
============

Since our `:timeout` option now handles redirects, we need to be able to
watch LiveViews that are redirected to.

To do that, we update the case when the watcher process tell the test
that the live view has redirected. In those cases, we used to perform
the action directly -- but that action is being performed without any
knowledge of waiting.

Now, the redirect message from the watcher process includes the amount
of timeout left, and the `with_timeout` helper is called again. Note
that sometimes there are redirects to static views, so we handle that
too.

Finally, we realized we had an error in our `timeout_ref` functionality.
We were creating our own `ref` and then using that ref for both pattern
matching the `:timeout` call (which worked), and for trying to cancel
the timer (which didn't work b/c it was the wrong ref). This commit
fixes that by using the timer ref and _not_ pattern matching the timeout
ref. It seems sufficient to use the view's pid since we're doing that
already.