Use LiveView.Channel to fetch async pids
What changed?
=============

We use `LiveView.Channel.async_pids/1` to query for async pids instead
of going through the client proxy.

The `render_async` code ends up calling the `Channel.async_pids/1`
function in the client proxy process. By going directly to the Channel
code, we're able to receive the `exit` signals that we need when the
process has died. Otherwise (with the current implementation) we only
receive exit signals from the client proxy process -- and we do not get
the correct signals.

This removes the need for us to know anything about the proxy, which
simplifies things, and it makes it so we mostly depend on the `Channel`
functions (`ping` elsewhere, and `async_pids` here).

Co-authored-by: Randy Coulman <randy@randycoulman.com>