Simplify timeout logic. Attempt faster checks.
What changed?
=============

The LiveViewWatcher and LiveViewTimeout logic is quite complicated
because our `timeout` option is trying to handle async processes AND any
redirects as a result of async operations (including `handle_info`
messages).

The first implementation originally attempted to try a more brute force
method (but still quite effective) of trying to perform the `action`
(e.g. `assert_has`) every X ms until we hit the timeout or the LiveView
redirects.

The problem with that was that we were hitting some race conditions
when:

1. We checked if LiveView redirected or died. It didn't.
2. We'd _assume_ the LiveView was alive and perform the action
3. The test would blow up because we were sending a message to a dead
   LiveView

The problem was that between 2 and 3, our LiveView had died.

Tracking async processes
-----------------------

Because of that, we introduced a more complex solution. In that, we
copied some code from LiveView to track async processes (used in
`render_async`) and notify us that the processes have finished.

That works pretty well... except when those async processes (think async
assigns) finish too fast! In those cases, by the time we try to monitor
the LiveView and its async processes, the async processes have already
finished. Thus, our tests wait the ENTIRE timeout (which can sometimes
be large).

Users have reported a problem (and I've seen it too) where a test is
taking a really long time, when it seems like it shouldn't b/c the async
operation is really fast. They wonder why their test with several
subsequent `assert_has(..., timeout: 500)` is taking 1500ms. It's
confusing, and we therefore try to fix that there.

Syncing with the LiveView
---------------------------

At some point in our current implementation, we introduced something of
a synchronization mechanism with LiveView to reduce race conditions
(which isn't bullet proof, but it seems to work well enough): before we
perform an `action`, we first ping the LiveView to see if it's alive
(via `LiveView.channel.ping/1`).

LiveView (or LiveViewTest) itself does this synchronization in its
codebase. By doing that ourselves, we're also able to catch exits from
the LiveView process if it's dead right as we're about to perform the
`action`.

Our new attempt
---------------

Having introduced that sync mechanism, and knowing that the async
processes can lead to unexpected long wait times in tests, we now turn
back to the more brute force--but much simpler implementation--of trying
to perform the `action` ever X ms.

That simplifies the watcher implementation, since we move all async
processes. We also move all timeout related logic to the
`LiveViewTimeout` module itself, further simplifying the watcher code.