Allow calling `visit/2` sequentially
Resolves #172

What changed?
==============

There are scenarios when library users want to call `visit/2` multiple
times sequentially. One example is the need to navigate to a different
page when a menu is in a `root` layout (and hence is not rendered by
LiveView).

This change makes it so we can call `visit/2` sequentially. It's also a
correct implementation of the `Drive` protocol for `Live` and `Static`.
Prior to this change, those were technically broken (and really, never
used).