Allow nesting of `within/3`
What changed?
============

This updates `within/3` to be able to nest `within/3` for both Live and
Static implementations.

Though not strictly required (since right now the last within always
wins), it can be confusing if someone nests `within/3` and it
unexpectedly finds something that would've been left out by the "outer"
within.

As part of this work, we create a `SessionHelpers` module to keep the
`within` implementation that was duplicated across `Live` and `Static`.
It's not the best module name, but it works for now. It might attract
more logic and get a better name in the future.