Refactoring: Make FormData a struct
What changed?
=============

We make `FormData` a struct so we can do two things:

- Encapsulate the underlying data structure (and later change it if we
  want to)
- To help add clarity to all of our form logic. Between Element.Form,
  ActiveForm, AttributesForm, FormPayload, and FormData, it can be
  confusing what each piece is doing for people who want to contribute
  to PhoenixTest. Making this an explicit struct makes it easier to
  understand the flow of code.

A good example of an improvement that came about b/c of the two points
above is the added clarity as to where the "deduplication" of form data
should happen. Really, the duplication of form data is an artifact of us
keeping form data in a list (in order to handle multiple selects).
FormPayload need not know anything about that, so we move the logic into
`FormData.to_list`.