Support role="button" for `click_button/2`
Resolves #140

What changed?
============

Sometimes people want to click on elements that aren't buttons. We
already support clicking on typical things like inputs with type submit
or type button.

This commit improves that so we can also click on any element with the
`role="button"`. That way, when people want to click on spans or divs,
they can do so by adding a "button" role.

There's more work towards making that span or div fully accessible
(hence, the encouragement in the docs to use `<button>` tags), but at
least it's a nudge in the right direction.

https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Reference/Roles/button_role