Add LiveViewTimeout test (& simplify LiveViewWathcer test)
What changed?
============

We add a LiveViewTimeout test. The functionality is already tested via
integration tests, but we want to have a finer-grained test to
(hopefully) see if we can identify any race conditions. Maybe
we'll add property-based tests to try to detect

For now, we just introduce the test. And while we do that, we realize
two things:

1. We can also simplify the LiveViewWatcher test with some of the same
   things we did in the LiveViewTimeout test
2. We update the private functions inside LiveViewTimeout to be
   consistent with the order of arguments (i.e. `action` and `timeout`
   swapped for no reason)