Adds property-based tests
What changed?
============

Adds property-based tests to LiveViewTimeout and LiveViewWatcher.

Since we introduced new logic in `with_timeout`, we wanted to make sure
that we could test them under various scenarios (e.g. different
timeouts, times to perform an action, etc.). So, we add a couple of
property-based tests for LiveViewTimeout and LiveViewWatcher.

It's possible we don't need all of them (LiveViewTimeout uses the
watcher, after all). But for now, we leave them.