Catch exit when timeout is `0`
What changed?
=============

There's a scenario (often shows as a race condition), where the
`LiveViewTimeout.with_timeout` function fails to catch an exit. It took
some property-based tests (which we do not include because they were
written to fail, but they did catch the race condition), to realize what
was happening:

- `handle_watched_messages_with_timeout` gets called.
- The `new_timeout` is set to `0`
- We do NOT receive a message from the LiveView, so we call
  `with_retry`
- The call in `with_retry` fails because the LiveView process has exited
- We call `retry_fun` (which is `handle_watched_messages_with_timeout`)
  expecting to go back into the `receive` flow, but since the
  `new_timeout` (step 2 above) is `0`, we fall into the other function
  head.
- We try to perform the `action`, but the LiveView is dead so the
  process exits.

To fix that, we simply have to catch exits when trying that last
`action`. We do that in this commit. (And we add a regular test to test
that).