Deprecate select with `:from` in favor of `:option` (#191)
What changed?
============

Our `select` form helper takes in the option text first, and then it
takes the label text in a `:from` argument.

That is confusing since every other form helper takes in the label first
(as a positional argument), and then takes the "other value" (when it
has one) as a keyword list.

We did it like that because it seemed to read somewhat naturally, and
because of previous art (that's how Capybara's API had it). But really,
the confusion it causes isn't worth the way it reads.

So, we deprecate the `:from` key in favor of passing the option text in`
:option`.

Thus, we have to change:

```diff
- |> select("Option 1", from: "Select Label")
+ |> select("Select Label", option: "Option 1")
```

For now, things don't break. We include a deprecation warning which will
hopefully help people make the upgrade. We also include some information
in the upgrade guides.

NOTE that this doesn't require Drivers to change yet. We'll follow up
with that once we fully remove the `:from` option.