`assert_path` handles query params with []
What changed?
============

We improve `assert_path` to handle query params that have lists.

There are many cases when we want to assert the query includes things
like `users[]=frodo&users[]=sam`.

The assertion would present those as a list in the map:

```elixir
assert_path("/some/path", query_params: %{users: ["frodo", "sam"]})
```

This commit makes it so that we can do that.