Adds asserting & refuting by `:value`
What changed?
============

We introduce a new `:value` option to `assert_has/3` and `refute_has/3`.

It allows us to make assertions about the value of an `<input>`,
`<option>`, etc. That's mostly helpful when we want to check the state
of an input.

The option cannot be combined with `:text` (we raise), and it isn't
compatible with `:exact` (the value is always expected to be exact).

Aside from that, all other options are allowed (e.g. `:count`, `:at`,
etc.)