Handles LiveView forms that remove elements
What changed?
============

Contrary to Static forms, LiveView forms can dynamically remove inputs,
selects, etc.

Right now our form submission assumes every input that has been filled
(or defaulted) should remain when we submit the form. But that doesn't
handle the case when the form changes as a result of user's behavior on
the page.

This commit fixes that by comparing the form's elements right before
submission with the data that we have gathered from user interactions.
If the form no longer has a field with that `name` on it, we remove it
from the form submission (assuming it's been removed one way or
another).