Adds: asserting by value and label (#204)
What changed?
=============

Now that we've introduced the ability to assert the value of an element
(i.e. a form field), we improve the `assert_has/3` and `refute_has/3` to
allow asserting by `label`.

The `:label` option only works with `:value`, not with `:text`. It makes
the options we can pass to `assert_has/3` and `refute_has/3` a little
strange--you now need to know the correct incantation of options to
pass--but it seems worth the trade off.

If things get any more complicated, it may be worth adding helpers like
`field/1` that would allow a more composable API (e.g.
`assert_has(field(value: "foo", label: "bar"))`). But for now, this
works as-is without having to introduce new helpers or modules.