Do not raise non-existent path if URL is external (#206)
What changed?
=============

Commit 668448ab introduced raising on paths that were not defined in the
router.

That provides a nicer experience when TDDing, but people are using
`assert_path` to assert they've landed on an external URL (think
authentication failing or logging out of the app).

In those cases, the request is a 404. But since `assert_path` is just
checking the `conn.request_path`, we're able to assert something about
the path even if PhoenixTest cannot actually visit that external URL.

This commit adds a check to see if the path we're trying to validate
with the router is internal or external.

- If external, we do not raise (since we might want to `assert_path`).
- If internal, we raise so that it's clear the path isn't defined in the
  application's router.

NOTE: to check if the request is external, we check the request's host
against two hosts -- the endpoint's host and the [default
`www.example.com` host] set in `Plug.Adapters.Test.Conn` (which gets
used by `Phoenix.ConnTest.build_conn/0`). If either host matches, then
the request is "internal".

[default host]: https://github.com/elixir-plug/plug/blob/1947edc8171f923ecc262efba6a0946264fa5582/lib/plug/adapters/test/conn.ex#L44