Rely on data in form for inputs_for (ordinal fields)
What changed?
=============

When a form uses index-based names for fields (like when using
`inputs_for`), and when it's possible to dynamically add/remove more of
those nested forms, it's possible for a form to remove a "middle" form
which causes later forms to shift their index down by 1.

But the problem is that PhoenixTest knows nothing about the removal. All
it can see is the HTML on the page and what the user has "filled in"
across form actions like `fill_in`, `select`, etc.

Example
-------

For example, a user may have a form with nested forms, each of which has
a single text field. Suppose they fill in "AAA" as index 0, add another
form, put "BBB" as index 1, add another form, and put "CCC" as index 2.

When they remove "BBB", "CCC" becomes index 1.

PhoenixTest knows that the user filled in index 0 with "AAA", 1 with
"BBB", and 2 with "CCC". But it doesn't know that index 1 was removed.

At the time of submission PhoenixTest looks at the current state of the
form and compares it to the data we know the user filled in. If the user
filled in anything that is no longer on the form, we assume some
deletion happen, and thus we do not submit that data.

But with index-based names, that's a problem. When we removed "BBB",
"CCC" became index 1, but PhoenixTest doesn't know that. All it sees is
that index 2 was removed but 1 is still on the page, so it removes "CCC"
from the "filled in" data and keeps "BBB".

The "fix"
--------

To fix that, we do not include any fields that have index-based fields
in our "filled in" data. Instead, we rely on whatever is present on the
page at the moment of submission for those fields.

That only works if LiveView is using "controlled inputs" (inputs whose
value LiveView fully controls), but that seems like a safe assumption to
make since most LiveView forms are probably using the helpers provided
by Phoenix (which set the `value` on `phx-change`).

In fact, it's possible that we could completely rely on the data on the
page at the time of submission for LiveViews (so long as the assumption
holds true). For now, we do not change it to that since many of our
tests use uncontrolled inputs.