Allow `JS.patch` in `phx-click` bindings
Resolves: #215

What changed?
============

We should allow `JS.patch` in `phx-click` bindings. By default, that
works with our `click_link`. But if you have that in a `click_button`,
we check the `phx-click` attribute first, and right now that fails.

This commit fixes that issue by adding `patch` to the list of allowed JS
commands we know LiveViewTest can handle.