Document simple form + ordinal `inputs_for` inputs
What changed?
==============

Commit f9d1bb9 had to revert the work done in d99c572 because it was
introducing a regression. The regression was not properly documented by
a test in this codebase but rather in a codebase that was using
phoenix_test. We don't want to have to rely on external codebases to
check for regressions.

This commit adds a simple test of a form that uses `inputs_for` with a
`phx-submit` but doesn't use a `phx-change`. This is the test that
should catch that regression when we (once again) attempt to handle
ordinal inputs that get deleted and added dynamically.

For more on using inputs_for, check out Phoenix's docs on dynamically
adding and removing inputs:

https://hexdocs.pm/phoenix_live_view/Phoenix.Component.html#inputs_for/1-dynamically-adding-and-removing-inputs