Revert "Rely on data in form for inputs_for (ordinal fields) (#208)"
What changed?
=============

This reverts commit d99c572e82939dc2e462ddab65df2d65238848fc.

Why?
----

The commit introduces a regression. Thankfully, we haven't shipped that
version of the code yet. So, it shouldn't cause breaking changes.

The regression comes from a LiveView form that only has a `phx-submit`
(no `phx-change`) but it uses ordinal inputs. Since the commit above
ignored our "active form" and there's no `phx-change` tracking the HTML
form (i.e. the inputs are "uncontroller"), then when the form is
submitted the data isn't present.

We hope to support dynamic addition/removal of fields, but for now, we
remove these changes.