Fix: allow multiple select and checkboxes
What changed?
=============

When we refactored `FormData`, we lost the ability to have multiple
selects and multiple checkboxes in some scenarios.

For example:

- when we have a default checkbox "checked" and add a new one
- when we use `select` multiple times in a row

This commit improves the merging behaviour of two FormData structs so
that we can handle key names that allow for multi-selects (i.e. they end
in `[]`)

Thanks to Fredrik for providing test cases for these failures.

Co-authored-by: Fredrik Teschke <f@ftes.de>