Use Floki 0.38 and fix deprecations (#241)
What changed?
=============

Floki has been warning for some time that we cannot pass strings to
`Floki` functions -- we should always `parse_document` or
`parse_fragment` first.

Well... the deprecations have caught up to us!

This commit updates PhoenixTest to handle that. It doesn't do it in the
cleanest way (i.e. ensuring we're always passing the correct value to
`PhoenixTest.Html`), but it seems like a pragmatic way.

Since most of PhoenixTest's interactions with Floki happen through
`PhoenixTest.Html`, we wrap those functions so that we parse values if
they're strings.