feat: handle uncheck on checkbox with phx-click and phx-value attributes
When a checkbox has a `phx-click` attribute, it correctly follows a flow
where that event is sent to the live view. The `check/3|4` functions
properly send along any `phx-value-` attributes in the params map to the
`handle_event/3` callback. `uncheck/3|4` would not do this, and would
override the params sent to the live view with an empty map.

The behavior of sending an empty params map seems expected for
checkboxes that are not in forms, but do have a `phx-click` attribute,
so this commit first handles the case where any `phx-value` attributes
are present. If none are present, it falls through to sending the empty
hash as before.