support phx-value and JS command values with uncheck (#246)
Closes: https://github.com/germsvel/phoenix_test/issues/237

Sorry of the delay, had to work through submission approval.

This PR updates `uncheck/3|4` to check if the checkbox has a `phx-click` attribute and either also has `phx-value` attributes or the `phx-click` attribute is a JS command with a value assigned, and makes sure those values are passed as params to the `handle_action/3` callback.

If there is a `phx-click` attribute but no identifiable values, the previous behavior of using an empty map for params is maintained for compatibility.