Preserve order of appearance for array inputs (#248)
Elements in lists/arrays in forms (referenced by unindexed field_name[]) have been aggregated in the order they appear in the HTML document before the refactoring in b1400b3
Since then, the order has been the opposite.
This works in most cases, but sometimes the order is significant and flipping it can become problematic.

This PR reverts the order to the "natural" order of occurrence in the document.