Add credo NoOpenBrowser check (#207)
Add a NoOpenBrowser credo check for those 
who want to use a credo check to ensure that
we do not leave an `open_browser` call. 

It only works when using a relative path in
requires, so we have to dig into `./deps/...`

There may be better ways to do that, but 
for now, we include this as is.