Fix: do not return empty potential matches in error
When changing to LazyHTML, the potential matches found in
`Query.find_one_of` can be empty nodes. The best way to handle is to
check for emptiness with `Enum.empty?`. So we do that here.

We also refactor the `Live.click_button` implementation so it doesn't
pass a `button: button` option, but rather, we call a
`handle_click_button` with a `button` consistently.