Fix: keep provided selector in `click_button/3`
What changed?
=============

Whenever we pass a selector in `click_button/3`, we're correctly using
the selector to find the button. But then we're incorrectly relying on
the parsed button's data to generate a new selector.

That happens because we're no longer using the provided selectors.
Instead, we use the parsed button's selector.

That's not typically a problem unless the selector provided had a
descendant selector (meaning it was using data outside of the button
itself). In those cases, subsequent attempts to use `button.selector`
fail if there are duplicates on the page.

To fix that, we simply rely on the selector provided so long as it's
something more complex than mere "button". If it's just "button", we
know that the selector we can generate from the parsed data is equal or
more specific. But if the user passes a selector, and we found a button
with that selector, we should trust that selector the to be a good
enough for the rest of the operations.