Refactor: adds PhoenixTest.Operation to keep rendered html
What changed?
=============

So far, anytime we need to get the rendered html, we call `render_html`.
Typically that only happpens once per operation (e.g. `click_button`,
`click_link`, etc.). But in some small instances, we might call it more
than once, simply because we didn't want to pass the `html` around.

Furthermore, buttons, fields, and selects kept the source html (from
which they were found) in a `source_raw` field. That's really not
something a `%Button{}` or `%Field{}` struct should need to know about.

In reality, we're missing an abstraction, and that's what we introduce
here. We add an `%Operation{}` struct that for now has a name (e.g.
`click_button`) and the `html` of the operation.

Thus, it serves as the operation's token.

So, we now have a session, a current form, and an operation -- three
distinct data-store structs that keep information at different levels.
We set the operation in all operations, even if we don't render html
(like in some `click_link`s) so that it's always correct when debugging
or looking at the session's `current_operation`.

NOTE: while doing this, we also removed the `raw` key from many structs
since those weren't used at all. It was a premature optimization,
introduced because we guessed we might need it in some future.