Refactor: set button.type when parsing
What changed?
=============

The previous commit added a check to see if a button is _NOT_ of
`type="button"` when we're evaluating it for a correct button for a form
submission.

It evaluated the `type` attribute on the fly. This commit moves that to
the `build` step, and we set the `button.type` as a new attribute in the
struct.