Improve `HTML.inner_text`. Rename `element_text` (#265)
What changed?
==============

The purpose of `HTML.inner_text` was always to try to extract the text
of an element without going too deeply into its children element. But to 
try to include those "extra" portions of text that might be in an `<em>` 
tag (or the like).

For example, we probably want to be able to capture "hello world" out of
`<div>hello <em>world</em></div>`, but we don't want to include text
inside of a `<textarea>` when we're just trying to match on the label's
text.

So, to handle all of that, we reverse the logic inside of `inner_text`
-- instead of excluding data for "select" and "textareas", we only
include additional text from children nodes when they are "known text
nodes".

We define "known text nodes" as those listed in [MDN Element docs] in
the [text content] or [inline text semantics] sections. And rename 
`Html.inner_text` to `Html.element_text` to try to capture the intent.

[MDN Element docs]: https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements
[text content]: https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements#inline_text_semantics
[inline text semantics]: https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements#inline_text_semantics