Add bin/create_changelog script
We add a bin/create_changelog script to help generate the changes for
the Changelog out of the commits since the last tagged release.

It includes the last tagged release -- might be we don't want that in
the future, but it's a nice visual confirmation that we got the correct
commit range.