feat: Target nested LiveViews with `within/2` (#270)
What changed?
=============

We allow targeting of nested LiveViews with `within/2`. Under the hood,
that will check if the HTML node the CSS selector is targeting is a
LiveView (we check by proxy by checking if it has a `data-phx-session`
attribute).

If it is a nested LiveView, we use LiveViewTest's `find_child_live` to
scope the operations inside `within/2` to that new view.

Why do we need that?
--------------------

Without finding the nested LiveView, we can interact with forms,
buttons, etc. but any event that gets trigered gets sent (incorrectly)
to the parent LiveView.