fix: Go back to excluding fields in `element_text`. Remove full LazyHTML.text (#273)
What changed?
=============

Commits 37bc521 (released in 0.8.3) and d19c6cf (released in 0.9.0)
tried to improve what was `HTML.inner_text` to be more coherent in text
parsing (i.e. what is `Html.element_text` vs `Html.text`). But in doing
so, we introduced regressions from the behavior we had in 0.8.2.

This commit goes back to behavior that more closely matches 0.8.2 by
making `Html.element_text` (what `inner_text` is now called) to always
include text inside nested tags (with the exception of `select` and
`textarea` if they aren't the top-level tag in the query). And we no
longer use `LazyHTML.text` at all -- since that doesnt' seem to exclude
newlines the same way we were doing so in our internal parsing.