feat: Raise an explicit error if target of `button_click` is disabled
I spent awhile trying to debug why a coworker's test was failing. The error being printed was:

```
** (ArgumentError) Expected element with selector "#my-button-id" and text
       "Preview" to have a valid `phx-click` attribute or belong to a `form` element.
```

This was confusing, because we could see the button did indeed belong to a form. What we were missing was that our test was poorly constructed, and the button was disabled.

This PR adds an explicit error for this case to make debugging easier.